+++ src/content/panorama/scripts/custom_game/game_setup/game_setup.js
@@ -1005,7 +1005,7 @@ function onBacktrackButton() {
 	var masterRoot = $.GetContextPanel();
     masterRoot.SetHasClass('phase_option_selection_selected', selectedPhase == PHASE_OPTION_SELECTION || reviewingOptions);
     masterRoot.SetHasClass('review_selection', reviewingOptions);
-    masterRoot.SetHasClass('phase_selection_selected', selectedPhase == PHASE_SELECTION && !reviewingOptions);
+    masterRoot.SetHasClass('phase_selection_selected', (selectedPhase == PHASE_SELECTION || selectedPhase == PHASE_INGAME) && !reviewingOptions);
 	
 	$('#backtrackBtnTxt').text = $.Localize((reviewingOptions)? 'reviewReturn': 'reviewOptions');
 }